% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{process_data}
\alias{process_data}
\title{Function \code{process_data}}
\usage{
process_data(filename)
}
\arguments{
\item{filename}{file name of csv data}
}
\value{
data frame
}
\description{
Process some data generated by \code{generate_data}.
}

