% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{wrapper_function}
\alias{wrapper_function}
\title{Function \code{wrapper_function}}
\usage{
wrapper_function()
}
\value{
result of gendata
}
\description{
Wrapper function to call generate_data. \code{wrapper_function}
tests if \code{remake} keeps track of changes to nested functions.
}

